Started of rebuilding the integration tests
Divided the controller tests for separate files.
Added new logic to CreateRecordTests for our custom responses.
