+++ src/Core/MoneyManager.Application/Functions/Records/Commands/CreateRecord/CreateRecordCommandValidator.cs
@@ -10,6 +10,8 @@ namespace MoneyManager.Application.Functions.Records
                 .NotEmpty()
                 .WithMessage("{PropertyName} is required")
                 .NotNull()
+                .MinimumLength(2)
+                .WithMessage("{PropertyName} must be at least 2 characters long")
                 .MaximumLength(50)
                 .WithMessage("{PropertyName} must not exceed 50 characters");
